changed from Java 7 Date/Time to Java 8 LocalDateTime. Added emtpy classes for tests
1: Major refactoring of all classes. I am now using LocalDateTime from Java SE 8 instead of using the deprecated methods from Java 7.

2: Added empty classes that will be fleshed out to tests.

All Classes have been updated to reflect changes and are now functioning.
Only a bare minumum of methods are implemented. Just enough to do some very rough testing through psvm.

Atm the Entries hold a correct time and date.
Entries are able to be added to the calendar.
The calendar can return a list of the entries added.

The class hierarchy appears to be functioning.

Emtpy test classes have been added. The path forward will be creating tests, then implementing.